research/requirements.txt
newspaper4k>=0.9.3
psutil>=5.9.0
requests>=2.28.0
python-dateutil>=2.8.2
argparse>=1.4.0
matplotlib>=3.5.0
numpy>=1.22.0
lxml[html_clean]>=5.3.0
nltk>=3.8.1
pandas>=2.0.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
pytest>=7.3.1
pytest-cov>=4.1.0 